Extract `discover` to new `llvm` package in tools
